requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
blinker==1.5
boto3==1.26.0
Django==4.1.3
dynamorm==0.11.0
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
marshmallow==3.18.0
djangorestframework==3.14.0
dropbox==11.36.0
pandas==1.5.1